src/requirements.txt
asgiref==3.2.7
astroid==2.4.0
autopep8==1.5.2
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
django-cors-headers==3.7.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.1.2
psycopg2==2.8.5
pycodestyle==2.5.0
PyJWT==2.0.1
pylint==2.5.0
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
urllib3==1.25.9
wrapt==1.12.1
